Proposal for buffered sensor/processor/capture functionality.
This is a proposal. This functionality needs to be better considered,
and possibly moved to ProcessorMotionBuffered.

Motion sensors/processors/captures can be of two types:

 - Incremental: each new data increments the state and the old history
is lost. Useful for slow sensors providing motion data, such as a laser
used with ICP to copute incdremental frame transforms.

 - Buffered: each new data is appended to a buffer waiting to be
integrated later. The integral can be reset and re-integrated at any
time by retaking the buffered data. Useful for quick sensors such as IMU
where data comes much faster than the solver loop, and also the
integration may be impossible to update if the initial condition change,
for example after a solver update.