Redesign of tracker for a flexible behavior with Landmarks
Not all trackers will use landmarks. Here we added a boolean to indicate
so, and check this condition for controlling the main algorithm of
process(). Basically, if no landmarks are used, the trackes will work
only at the Capture and Features level, establishing constraints between
origin and last Features. If landmarks are used, one can implement
active search strategies, create new landmarks as they are discovered,
and establish the constraints between Features and Landmarks.