Implement split(), first version (look longer note...)
The current versions returns in a new buffer the part older than the
given time stamp.

Possibly we want the opposite: keep the old part in the current Capture, 
and close this Capture in a new keyframe. Then return the new part in a
separate buffer to provide it to the non-key Capture to continuer
integrating there.